types: Give propertyDeclaration a proper argument type
The `node.parameters` case turns out to be impossible: that property
doesn't exist on either `ts.VariableDeclaration` or
`ts.PropertyDeclaration`.

And these are the only two call sites for this `propertyDeclaration`
function, so the argument really is always of one of those types --
it's not that there's another call site somewhere passing in something
else that's concealed by an `any`.

Also use proper `ts.isFoo(node)` guards to simplify the logic.