printers [nfc]: Correct the type that identifiers.print() returns
The type annotation said it returns string, but in fact it can also
return a function.

At one call site, this resolves a ts-expect-error, because we were
invoking the function when it is one.  But at two others, it means we
need to introduce one -- they sure look like they'd blow up if they
actually got a function there.  I'm not sure whether those are live
bugs, or if there's a subtle reason they're OK; I experimented briefly
and didn't find a way to trigger that.

Also initialize the `identifiers` global directly with an object
literal, rather than starting with `Object.create(null)`, because
that lets us write down a type annotation for it.  This just requires
us to be a bit more careful when looking up a key there.