printers: Fix interface-extends to stay explicitly inexact when `inexact` false
These object types must be inexact in order for the intersection to
have the intended effect.

When the `inexact` option was false, this logic would end up adjusting
part of the syntax as if we wanted an exact object type.  The net
result was an object type that was ambiguously exact or inexact, with
no `{|` and `|}` but no `...` either.

A bonus is that when the interface has just one member of its own, the
formatting is nicely more compact.  This causes a number of snapshots
to update.

Also add a test, with `inexact: false`.