printers: Drop now-redundant comments on enum members
Originally the implementation of enum members was rather indirect, and
these comments provided useful information about the original value.
But since a2cd5f01d which simplified the implementation, the value
shown earlier on the same line has had exactly the same content as the
comment.  So we can now drop the comment without losing anything.