printers: Use common printDefaultObjectType for type literals
This produces exactly the same types: the existing code was already
successfully producing an explicit inexact object type when the
`inexact` option is true, and an explicit exact object type when
that option is false.

The formatting in the case of a type with a single member is nicely
more compact, though, and some snapshots change as a result.