printers [nfc]: Add methods printInexactObjectType and friends
These will let us replace a number of separate implementations at
different spots in our code of printing a Flow object type.

In particular, these will always emit a form that's explicitly either
exact or inexact, eliminating places where we currently emit an
ambiguous form like `{ x: number }`.  See Flow docs:
  https://flow.org/en/docs/types/objects/#toc-explicit-inexact-object-types

By my count, we currently have 5 places where we emit this form and
intend it to mean an inexact type, and 2 where we intend the opposite.

These will also let our own code be explicit about whether it
specifically intends an exact object type, or an inexact object type,
or wants the exactness to be controlled by the `inexact` option.
At least one of those bugs (the one affecting an `interface` with an
`extends` clause) appears to be due to confusion where one bit of code
is trying to specifically produce an inexact object type, but another
bit that it's calling is trying to let the exactness be controlled by
the `inexact` option.  I suspect that sort of bug will be easier to
avoid when the choice is explicit.